+++ Empregado.hpp
@@ -7,21 +7,14 @@
 class Empregado {
 	
   public:
+    // Atributos
     double salarioHora;  
     double quotaMensalVendas;  
 
+    // Metodos
+    double pagamentoMes( double horasTrabalhadas );
 
-    double pagamentoMes(double horasTrabalhadas) {
- 
-      double t = horasTrabalhadas;
-	  
-	  //Cálculo de hora extra (+50% se horasTrabalhadas > 8)
-      if (horasTrabalhadas > 8) {
-        double x = horasTrabalhadas - 8;
-        t += x / 2;
-      }
-	  return t * salarioHora;
-    }
+
 	
 };
