+++ Vendedor.cpp
@@ -1,14 +1,7 @@
 #include <string>
-#include "Empregado.hpp"
+#include "Vendedor.hpp"
 
-class Vendedor : public Empregado {
-
-  public:
-    std::string nome;  	  
 	
-	double quotaTotalAnual() {
-	  return quotaMensalVendas * 12;
-	}
-	
-};
-
+double Vendedor::quotaTotalAnual() {
+    return quotaMensalVendas * 12;
+}
