+++ Vendedor.hpp
@@ -1,0 +1,12 @@
+#include <string>
+#include "Empregado.hpp"
+
+class Vendedor : public Empregado {
+
+  public:
+    std::string nome;  	  
+	
+	double quotaTotalAnual();
+	
+};
+
