+++ main.cpp
@@ -1,5 +1,7 @@
-#include "Engenheiro.cpp"
-#include "Vendedor.cpp"
+/*** Eduardo Capanema ***/
+#include "Engenheiro.hpp"
+#include "Vendedor.hpp"
+/*** 2020041515 ***/
 
 int main() {
 
@@ -8,7 +10,7 @@ int main() {
   eng1.salarioHora = 35;
   eng1.projetos = 3; 
   std::cout << "Nome: " << eng1.nome << std::endl;
-  std::cout << "Salario Mes: " << eng1.pagamentoMes(9.5) << std::endl;
+  std::cout << "Salario Mes: " << eng1.pagamentoMes( 9.5 ) << std::endl;
   std::cout << "Projetos: " << eng1.projetos << std::endl;
   std::cout << std::endl;
   
@@ -17,7 +19,7 @@ int main() {
   eng2.salarioHora = 30;
   eng2.projetos = 1; 
   std::cout << "Nome: " << eng2.nome << std::endl;
-  std::cout << "Salario Mes: " << eng2.pagamentoMes(8) << std::endl;
+  std::cout << "Salario Mes: " << eng2.pagamentoMes( 8 ) << std::endl;
   std::cout << "Projetos: " << eng2.projetos << std::endl;  
   std::cout << std::endl;
   
@@ -26,7 +28,7 @@ int main() {
   eng3.salarioHora = 30;
   eng3.projetos = 2; 
   std::cout << "Nome: " << eng3.nome << std::endl;
-  std::cout << "Salario Mes: " << eng3.pagamentoMes(8) << std::endl;
+  std::cout << "Salario Mes: " << eng3.pagamentoMes( 8 ) << std::endl;
   std::cout << "Projetos: " << eng3.projetos << std::endl;  
   std::cout << std::endl;
   
@@ -37,7 +39,7 @@ int main() {
   vend1.quotaMensalVendas = 5000;
   
   std::cout << "Nome: " << vend1.nome << std::endl;
-  std::cout << "Salario Mes: " << vend1.pagamentoMes(6) << std::endl;  
+  std::cout << "Salario Mes: " << vend1.pagamentoMes( 6 ) << std::endl;  
   std::cout << "Quota vendas: " << vend1.quotaTotalAnual() << std::endl;
   std::cout << std::endl;
   
@@ -47,7 +49,7 @@ int main() {
   vend2.quotaMensalVendas = 3000;
   
   std::cout << "Nome: " << vend2.nome << std::endl;
-  std::cout << "Salario Mes: " << vend2.pagamentoMes(8) << std::endl;  
+  std::cout << "Salario Mes: " << vend2.pagamentoMes( 8 ) << std::endl;  
   std::cout << "Quota vendas: " << vend2.quotaTotalAnual() << std::endl; 
   std::cout << std::endl;  
 	
@@ -57,8 +59,9 @@ int main() {
   vend3.quotaMensalVendas = 4000;
   
   std::cout << "Nome: " << vend3.nome << std::endl;
-  std::cout << "Salario Mes: " << vend3.pagamentoMes(8) << std::endl;  
+  std::cout << "Salario Mes: " << vend3.pagamentoMes( 8 ) << std::endl;  
   std::cout << "Quota vendas: " << vend3.quotaTotalAnual() << std::endl;  
   
-  return 0;	
+  return 0;
+
 }
